[update]patch&delete-test-add-only-owner-using

+++ app/Http/Controllers/ContactsController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Resources\Contact as ContactResources;
 use App\Models\Contact;
 use Illuminate\Http\Request;
 
@@ -9,32 +10,36 @@ class ContactsController extends Controller
 {
     public function index()
     {
-        //dd(request()->user()->contacts);
-        return request()->user()->contacts;
+        $this->authorize('viewAny', Contact::class);
+
+        return ContactResources::collection(request()->user()->contacts);
     }
 
     public function store()
     {
+        $this->authorize('create', Contact::class);
+
         request()->user()->contacts()->create(array_merge(['user_id' => request()->user()->id], $this->validateData()));
     }
 
     public function show(Contact $contact)
     {
-        // ２つのモデルが異なるかを判定する
-        if(request()->user()->isNot($contact->user)) {
-            return response([], 403);
-        }
+        $this->authorize('view', $contact);
 
-        return $contact;
+        return new ContactResources($contact);
     }
 
     public function update(Contact $contact)
     {
+        $this->authorize('update', $contact);
+
         $contact->update($this->validateData());
     }
 
     public function destroy(Contact $contact)
     {
+        $this->authorize('delete', $contact);
+
         $contact->delete();
     }
